Add files via upload
Here we can find all the files the project is consisted of. It's a mix of documents where the most important ones are "rob_sic", which is the main code which reunites all the points of the project, especified in the document "ElaboratoCdR18_19.pdf". Apart from ".m" files we can see there are various Simulink projects where all the types of control are built (robust and adaptative).
Finally, the main document to present to the professor is "PROGETTO DE CORSO_C.pdf", written in italian because this project has been made during an Erasmus program in Naples, Italy.